requirements.txt
pillow>=10.0.0
openai>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0
streamlit>=1.32.0
